# backend/openapi-nodegen-api-file.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: openweathermap-swagger
  description: A sample API
  contact:
    name: Swagger API Team
    email: john@boats.io
    url: https://github.com/johndcarmichael/boats/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
tags:
  - name: weather
    description: Weather data
paths:
  /weather/:
    get:
      tags:
        - weather
      summary: Get weather
      description: Query with varying content
      operationId: weatherGet
      parameters:
        - $ref: '#/components/parameters/queryAppId'
        - $ref: '#/components/parameters/queryCountry'
        - $ref: '#/components/parameters/queryId'
        - $ref: '#/components/parameters/queryLat'
        - $ref: '#/components/parameters/queryLon'
        - $ref: '#/components/parameters/headerSearchId'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Weather'
        '400':
          description: Invalid Identifier supplied
        '404':
          description: Round not found
      security:
        - apiKeyAdmin: []
        - jwtToken: []
      x-passRequest: true
  '/weather/{id}':
    get:
      tags:
        - weather
      summary: Get weather
      description: Query with varying content
      operationId: weatherIdGet
      x-permission: getWeatherDetail
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherFull'
        '400':
          description: Invalid Identifier supplied
        '404':
          description: Round not found
      security:
        - apiKeyAdmin: []
        - jwtToken: []
      x-passThruWithoutJWT: true
  # /rain/:
  #   get:
  #     tags:
  #       - weather
  #     summary: Get weather
  #     description: Query with varying content
  #     operationId: rainGet
  #     parameters:
  #       - $ref: '#/components/parameters/queryCountry'
  #     responses:
  #       '200':
  #         description: successful operation
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Weather'
  #       '400':
  #         description: Invalid Identifier supplied
  #       '404':
  #         description: Round not found
  #   post:
  #     tags:
  #       - weather
  #     summary: Get weather
  #     description: Query with varying content
  #     operationId: rainPost
  #     requestBody:
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/Coord'
  #       required: true
  #     responses:
  #       '200':
  #         description: successful operation
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Weather'
  #       '400':
  #         description: Invalid Identifier supplied
  #       '404':
  #         description: Round not found
  #     x-passRequest: true
  #   put:
  #     tags:
  #       - weather
  #     summary: Get weather
  #     description: Query with varying content
  #     operationId: rainPut
  #     requestBody:
  #       content:
  #         application/x-www-form-urlencoded:
  #           schema:
  #             type: object
  #             properties:
  #               image:
  #                 description: An image to upload
  #                 type: string
  #                 format: binary
  #             required:
  #               - image
  #     responses:
  #       '200':
  #         description: successful operation
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Weather'
  #       '400':
  #         description: Invalid Identifier supplied
  #       '404':
  #         description: Round not found
  #     x-passRequest: true
servers:
  - url: https://samples.openweathermap.org/data/2.5
  - url: https://api.example.com
  - url: https://api.example.com:8443/v1/reports
  - url: http://localhost:3025/v1
  - url: http://10.0.81.36/v1
  - url: ws://api.example.com/v1
  - url: wss://api.example.com/v1
  - url: /v1/reports
  - url: /
  - url: //api.example.com
components:
  parameters:
    pathId:
      in: path
      name: id
      required: true
      schema:
        type: string
    queryAppId:
      in: query
      name: appid
      required: true
      description: |
        Grab a free API key from here: https://openweathermap.org/price
      schema:
        type: string
    queryCountry:
      in: query
      name: q
      required: false
      description: >
        q city name and country code divided by comma, use ISO 3166 country
        codes. for example: q=London,uk or just q=London
      schema:
        type: string
    queryId:
      in: query
      name: id
      required: false
      description: >
        City ID based on city.list.json.gz can be downloaded here
        http://bulk.openweathermap.org/sample/
      schema:
        type: number
    queryLat:
      in: query
      name: lat
      required: false
      description: |
        Needs to be paired with lon
      schema:
        type: number
    queryLon:
      in: query
      name: lon
      required: false
      description: |
        Needs to be paired with lat
      schema:
        type: number
    headerSearchId:
      in: header
      name: Search-Id
      description: Unique search id
      example: 569eecd9-9962-4aed-a0f0-30476c6a82ed
      schema:
        type: string
        pattern: ^Bearer .+$
  securitySchemes:
    jwtToken:
      type: apiKey
      name: Authorization
      in: header
    apiKeyAdmin:
      type: apiKey
      name: x-api-key
      in: header
    apiToken:
      type: http
      scheme: bearer
    basic:
      type: http
      scheme: basic
  schemas:
    AdditionalProps:
      type: object
      additionalProperties:
        type: string
    ArrayProps:
      type: array
      items:
        type: string
    Coord:
      type: object
      properties:
        lon:
          type: number
        lat:
          type: number
    JwtAccess:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    WeatherFull:
      allOf:
        - $ref: '#/components/schemas/Weather'
        - type: object
          required:
            - name
          properties:
            id:
              type: string
            name:
              type: string
    Weather:
      type: object
      required:
        - coord
      properties:
        coord:
          $ref: '#/components/schemas/Coord'
        weather:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              main:
                type: string
              description:
                type: string
              icon:
                type: string
        base:
          type: string
        main:
          type: object
          properties:
            temp:
              type: number
            pressure:
              type: number
            humidity:
              type: number
            temp_min:
              type: number
            temp_max:
              type: number
            sea_level:
              type: number
            grnd_level:
              type: number
        wind:
          type: object
          properties:
            speed:
              type: number
            deg:
              type: number
            clouds:
              type: object
              properties:
                all:
                  type: number
        dt:
          type: number
        sys:
          type: object
          properties:
            message:
              type: number
            country:
              type: string
            sunrise:
              type: number
            sunset:
              type: number
        id:
          type: number
        name:
          type: string
        cod:
          type: number
    WeatherUpload:
      type: object
      properties:
        rain:
          type: boolean
        pdfAboutHowMuchBetterWinterIs:
          type: string
          format: binary
        startDate:
          type: string
          format: date